==> Notebooks_and_Skripts/01_Data_loading_cleaning_exporting.ipynb <==
import pandas as pd
from datetime import datetime

ars = pd.read_csv("../data/1429_1.csv" , low_memory=False)
ars.info()

ars.nunique()

ars.columns = [c.replace("reviews.","") for c in list(ars.columns)]
ars.columns

ars = ars [[ 'name', 'date', 'rating', 'text', 'title', 'doRecommend', 'username']]
ars.rename(columns={"name"        : "Product_name" ,
                    "date"        : "date_dirty",
                    "rating"      : "rating" ,
                    "text"        : "rating_text",
                    "title"       : "rating_title",
                    "doRecommend" : "recommended" ,
                    "username"    : "username"
                   } , inplace=True);

ars.Product_name = ars.Product_name.str.replace(",","")
ars.Product_name = ars.Product_name.str.replace('"','')
ars.Product_name = ars.Product_name.str.replace("'","´")
ars.Product_name = ars.Product_name.str.replace("\r","")
ars.Product_name = ars.Product_name.str.strip()

ars.rating_text  = ars.rating_text.str.replace(",","")
ars.rating_text  = ars.rating_text.str.replace('"','')
ars.rating_text  = ars.rating_text.str.replace("'","´")

ars.rating_title = ars.rating_title.str.replace(",","")
ars.rating_title = ars.rating_title.str.replace('"','')
ars.rating_title = ars.rating_title.str.replace("'","´")

ars.isnull().sum()

ars = ars.dropna(how='any')
ars.drop_duplicates(inplace=True)

ars.Product_name = ars.Product_name.apply(lambda x: str(x).split("\n",1)[0])

ars["date"]  = ars.date_dirty.apply(lambda x: datetime.fromisoformat(x[:10]));
ars.drop("date_dirty",axis='columns',inplace=True);

ars["year"]  = pd.DatetimeIndex(ars.date).year;
ars["month"] = pd.DatetimeIndex(ars.date).month;
ars["day"]   = pd.DatetimeIndex(ars.date).day;

ars.recommended = ars.recommended.astype('bool')

ars.head()

ars.shape

ars.info()

ars.nunique()

ars.to_csv("../data/amazon_reviews_clean.csv",index=False,encoding='utf-8')

max(ars.Product_name.str.len())

max(ars.rating_text.str.len())

max(ars.rating_title.str.len())

max(ars.username.str.len())

ars[ars.duplicated(subset="rating_text", keep=False)==True]
