==> Notebooks_and_Skripts/03_Feeding_the_SQL_DataBase.ipynb <==
## 1. Importamos librerías

import pandas as pd
import json

## 2. Importamos funciones para introducir ratings, users y productos en la BD en SQL.

from config.sql_tools import check
from config.sql_tools import insert_user
from config.sql_tools import insert_product
from config.sql_tools import insert_rating

Los tokens y la conexión a BD SQL está definida en el archivo `<config.configuration>`. 
Dentro de `<config.sql_tools>`, al principio de todo se importa la funcion `<config.configuration>` para poder acceder a la base de datos.


## 3. Cargamos el CSV ya limpio

ars = pd.read_csv("../data/amazon_reviews_clean.csv")

ars.info()

ars.head()

## 4. Iteramos fila por fila en todo el dataframe y para meter todas las filas en la BBDD

ratings_que_dan_error = []
for index,fila in ars.iterrows() :
    #print(index)
    if index%100 == 0:
        print(f"Se han cargado ya {index} de un total de {len(ars)}")
    try: 
        insert_rating(fila);
    except Exception as err:
        print(f"Row {index} da error. Se registra.     --> {type(err).__name__} : {err}")
        ratings_que_dan_error.append( {"row" : index , "error_type" : type(err).__name__ , "error_message" : err.args } )

len(ratings_que_dan_error)

ratings_que_dan_error

ratings_error_json = json.dumps(ratings_que_dan_error)

ratings_error_json

with open("../data/rows_generating_erros.json", "w") as f:
   json.dump(ratings_que_dan_error, f)
