==> Notebooks_and_Skripts/04_Api_calls.ipynb <==
# 1. Importamos librerías necesarias

import requests
import json
import pandas as pd

# 2. Comprobamos que todos los endpoints devuelven respuesta 200

    /                   [GET]
    /user_list          [GET]
    /product_list       [GET]
    /ratingsbyuser      [GET]
    /ratingsbyproduct   [GET]
    /newuser            [POST]
    /newproduct         [POST]
    /newrating          [POST]

requests.get("http://127.0.0.1:5000/")

requests.get("http://127.0.0.1:5000/user_list")

requests.get("http://127.0.0.1:5000/product_list")

usuario  = {"user"   : "RichBo"}
requests.get("http://127.0.0.1:5000/ratingsbyuser",data=usuario)

producto  = {"product"   : "Amazon Kindle Fire 5ft USB to Micro-USB Cable (works with most Micro-USB Tablets)"}
requests.get("http://127.0.0.1:5000/ratingsbyproduct",data=producto)

nuevo_usuario  = {"new_user"   : "USUARIO DE PRUEBA"}
requests.post("http://127.0.0.1:5000/newuser", data=nuevo_usuario)

nuevo_producto = {"new_product": "PRODUCTO DE PRUEBA" }
requests.post("http://127.0.0.1:5000/newproduct", data=nuevo_producto)

nuevo_rating = {"rating"        : 3                 ,
                "rating_text"  : "RATING DE PRUEBA" ,
                "rating_title" : "PRUEBA"           ,
                "date"         : "2021-11-29"       ,
                "recommended"  : True               ,
                "Product_name" : "PRUEBA_RATING"    ,
                "username"     : "PRUEBA_RATING"    }
    
requests.post("http://127.0.0.1:5000/newrating", data=nuevo_rating)

# 3. Comprobamos el contenido de la respuesta de cada Endpoint

    /                   [GET]
    /user_list          [GET]
    /product_list       [GET]
    /ratingsbyuser      [GET]
    /ratingsbyproduct   [GET]
    /newuser            [POST]
    /newproduct         [POST]
    /newrating          [POST]

requests.get("http://127.0.0.1:5000/").content

#requests.get("http://127.0.0.1:5000/user_list").content
#lo tapamos porque son 25.000 usuarios.

requests.get("http://127.0.0.1:5000/product_list").content

usuario  = {"user"   : "RichBo"}
requests.get("http://127.0.0.1:5000/ratingsbyuser",data=usuario).content

producto  = {"product"   : "Amazon Kindle Fire 5ft USB to Micro-USB Cable (works with most Micro-USB Tablets)"}
requests.get("http://127.0.0.1:5000/ratingsbyproduct",data=producto).content

nuevo_usuario  = {"new_user"   : "USUARIO DE PRUEBA"}
requests.post("http://127.0.0.1:5000/newuser", data=nuevo_usuario).content
#En este caso, el Endpoint nos devuelve el ID del nuevo usuario en la BBDD

nuevo_producto = {"new_product": "PRODUCTO DE PRUEBA" }
requests.post("http://127.0.0.1:5000/newproduct", data=nuevo_producto).content
#En este caso, el Endpoint nos devuelve el ID del nuevo producto en la BBDD

nuevo_rating = {"rating"        : 3                 ,
                "rating_text"  : "RATING DE PRUEBA" ,
                "rating_title" : "PRUEBA"           ,
                "date"         : "2021-11-29"       ,
                "recommended"  : True               ,
                "Product_name" : "PRUEBA_RATING"    ,
                "username"     : "PRUEBA_RATING"    }
    
requests.post("http://127.0.0.1:5000/newrating", data=nuevo_rating).content
#En este caso, el Endpoint nos devuelve el ID del nuevo rating en la BBDD

# 4. JSON-izamos las respuesta de cada Endpoint

    /                   [GET]
    /user_list          [GET]
    /product_list       [GET]
    /ratingsbyuser      [GET]
    /ratingsbyproduct   [GET]
    /newuser            [POST]
    /newproduct         [POST]
    /newrating          [POST]

#requests.get("http://127.0.0.1:5000/user_list").json()
#lo tapamos porque son 25.000 usuarios.

requests.get("http://127.0.0.1:5000/product_list").json()

usuario  = {"user"   : "RichBo"}
requests.get("http://127.0.0.1:5000/ratingsbyuser",data=usuario).json()

producto  = {"product"   : "Amazon Kindle Fire 5ft USB to Micro-USB Cable (works with most Micro-USB Tablets)"}
requests.get("http://127.0.0.1:5000/ratingsbyproduct",data=producto).json()

nuevo_usuario  = {"new_user"   : "USUARIO DE PRUEBA"}
requests.post("http://127.0.0.1:5000/newuser", data=nuevo_usuario).json()
#En este caso, el Endpoint nos devuelve el ID del nuevo usuario en la BBDD

nuevo_producto = {"new_product": "PRODUCTO DE PRUEBA" }
requests.post("http://127.0.0.1:5000/newproduct", data=nuevo_producto).json()
#En este caso, el Endpoint nos devuelve el ID del nuevo producto en la BBDD

nuevo_rating = {"rating"        : 3                 ,
                "rating_text"  : "RATING DE PRUEBA" ,
                "rating_title" : "PRUEBA"           ,
                "date"         : "2021-11-29"       ,
                "recommended"  : True               ,
                "Product_name" : "PRUEBA_RATING"    ,
                "username"     : "PRUEBA_RATING"    }
    
requests.post("http://127.0.0.1:5000/newrating", data=nuevo_rating).json()
#En este caso, el Endpoint nos devuelve el ID del nuevo rating en la BBDD

# 4. Pasamos a DataFrame los JSONs de Ratings


    /ratingsbyuser      [GET]
    /ratingsbyproduct   [GET]


usuario  = {"user"   : "RichBo"}
pd.DataFrame(requests.get("http://127.0.0.1:5000/ratingsbyuser",data=usuario).json())

producto  = {"product"   : "Amazon Kindle Fire 5ft USB to Micro-USB Cable (works with most Micro-USB Tablets)"}
pd.DataFrame(requests.get("http://127.0.0.1:5000/ratingsbyproduct",data=producto).json())
